refactor: web3modal
