+++ src/hooks/web3modal/index.ts
@@ -1,0 +1,1 @@
+export { promptAndInjectProjectId } from './promptAndInjectProjectId'
