+++ templates/vite-react/connectkit/_meta.ts
@@ -4,7 +4,6 @@ import { injectProviders } from '../../../src/hooks/connectkit'
 import { createTemplate } from '../../../src/utils'
 
 export default createTemplate({
-  default: true,
   title: 'ConnectKit',
   description: 'Vite (React) wagmi project with ConnectKit included',
   hooks: compose([
