--- templates/vite-react/web3modal/src/components/NetworkSwitcher.tsx
@@ -1,33 +1,0 @@
-import { useNetwork, useSwitchNetwork } from 'wagmi'
-
-export function NetworkSwitcher() {
-  const { chain } = useNetwork()
-  const { chains, error, isLoading, pendingChainId, switchNetwork } =
-    useSwitchNetwork()
-
-  if (!chain) return null
-
-  return (
-    <div>
-      <div>
-        Connected to {chain?.name ?? chain?.id}
-        {chain?.unsupported && ' (unsupported)'}
-      </div>
-
-      {switchNetwork && (
-        <div>
-          {chains.map((x) =>
-            x.id === chain?.id ? null : (
-              <button key={x.id} onClick={() => switchNetwork(x.id)}>
-                {x.name}
-                {isLoading && x.id === pendingChainId && ' (switching)'}
-              </button>
-            ),
-          )}
-        </div>
-      )}
-
-      <div>{error && error.message}</div>
-    </div>
-  )
-}
